pyproject.toml
[project]
name = "my-example-server"
version = "0.1.0"
description = "An example MCP server"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "gumstack-mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
server = "my_example_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["my_example_server"]

[tool.uv]
package = true